Add Near.block
